move aioeapi patch
