+++ pyproject.toml
@@ -161,8 +161,6 @@ strict_optional = true
 # for strict mypy: (this is the tricky one :-))
 disallow_untyped_defs = true
 
-mypy_path = "anta"
-
 [tool.pydantic-mypy]
 init_forbid_extra = true
 init_typed = true
